requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
gunicorn==21.2.0
pandas==2.1.0
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
python-dotenv==1.0.0